requirements.txt
pandas==2.2.2
numpy==2.1.1
scikit-learn==1.5.2
umap-learn==0.5.6
matplotlib==3.9.2
seaborn==0.13.2
scipy==1.13.1
pyyaml==6.0.2
